refactor: move to `QuickAdd::QuickAddNotSupported`
