+++ app/cleanup/quick_add/from_hardcoded_name_factory.rb
@@ -10,33 +10,15 @@ module QuickAdd
         QuickAdd::FromHardcodedSubgenusName.new(name_string)
       else
         if epithet? name_string
-          QuickAddRankNotSupported.new("Identified as 'epithet' (not supported)")
+          QuickAdd::QuickAddNotSupported.new("Identified as 'epithet' (not supported)")
         else
-          QuickAddRankNotSupported.new("Identified as #{identified_name_class.name} (not supported)")
+          QuickAdd::QuickAddNotSupported.new("Identified as #{identified_name_class.name} (not supported)")
         end
       end
     end
 
     def self.epithet? name_string
       name_string.split.size == 1 && name_string.downcase == name_string
-    end
-
-    class QuickAddRankNotSupported
-      def initialize error_message
-        @error_message = error_message
-      end
-
-      def can_add?
-        false
-      end
-
-      def synopsis
-        error_message
-      end
-
-      private
-
-        attr_reader :error_message
     end
   end
 end
