+++ app/cleanup/quick_add/quick_add_not_supported.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+module QuickAdd
+  class QuickAddNotSupported
+    def initialize error_message
+      @error_message = error_message
+    end
+
+    def can_add?
+      false
+    end
+
+    def synopsis
+      error_message
+    end
+
+    private
+
+      attr_reader :error_message
+  end
+end
